@@ -58,39 +52,83 @@
     ]
   },
   "jest": {
-    "preset": "jest-expo"
+    "preset": "react-native"
   },
   "dependencies": {
     "@absinthe/socket": "^0.1.10",
     "@absinthe/socket-apollo-link": "^0.1.11",
-    "@expo/vector-icons": "^6.3.1",
+    "@babel/polyfill": "^7.4.3",
     "@jumpn/utils-graphql": "^0.5.3",
     "@richeterre/apollo-upload-client": "^8.0.0",
+    "@types/react-redux": "^6.0.10",
+    "add": "^2.0.6",
     "apollo-cache-inmemory": "^1.2.5",
+    "apollo-cache-persist": "^0.1.1",
     "apollo-client": "^2.3.5",
     "apollo-link-context": "^1.0.8",
     "apollo-link-error": "^1.1.0",
     "apollo-link-http": "^1.5.4",
     "apollo-link-state": "^0.4.1",
+    "appcenter": "^1.10.0",
+    "appcenter-analytics": "^1.10.0",
+    "appcenter-crashes": "^1.10.0",
     "create-react-native-app": "^1.0.0",
-    "expo": "^25.0.0",
+    "es6-symbol": "^3.1.1",
     "global": "^4.3.2",
     "graphql-tag": "^2.9.2",
     "humps": "^2.0.1",
-    "native-base": "^2.4.3",
+    "javascript-state-machine": "^3.1.0",
+    "js-search": "^1.4.2",
+    "moment": "^2.22.2",
+    "native-base": "^2.8.1",
     "phoenix": "^1.3.0",
     "prop-types": "^15.6.1",
-    "react": "16.2.0",
-    "react-apollo": "^2.1.9",
-    "react-native": "0.52.0",
+    "react": "16.6.0-alpha.8af6728",
+    "react-apollo": "2.3.2-beta.4",
+    "react-native": "0.57.3",
+    "react-native-animatable": "^1.3.0",
+    "react-native-aws3": "^0.0.8",
+    "react-native-cached-image": "^1.4.3",
+    "react-native-calendars": "^1.26.0",
+    "react-native-chart-kit": "^2.4.0",
+    "react-native-code-push": "^5.5.2",
     "react-native-collapsible": "^0.11.2",
+    "react-native-collapsible-header": "^0.0.14",
+    "react-native-config": "^0.11.7",
+    "react-native-credit-card-input": "^0.4.1",
     "react-native-datepicker": "^1.6.0",
+    "react-native-device-info": "^0.24.3",
+    "react-native-document-picker": "^2.2.0",
+    "react-native-fetch-blob": "^0.10.8",
+    "react-native-firebase": "^5.2.3",
     "react-native-google-places-autocomplete": "^1.3.6",
+    "react-native-image-crop-picker": "^0.21.3",
+    "react-native-keyboard-aware-scroll-view": "^0.8.0",
+    "react-native-lightbox": "^0.8.0",
     "react-native-loading-spinner-overlay": "^0.5.2",
-    "react-native-modal-popover": "^0.0.10",
+    "react-native-onesignal": "^3.2.12",
+    "react-native-progress": "^3.5.0",
+    "react-native-rave": "https://github.com/gbenga504/rave-react-native.git",
+    "react-native-sentry": "^0.40.2",
     "react-native-slider": "^0.11.0",
+    "react-native-splash-screen": "^3.1.1",
     "react-native-star-rating": "^1.0.9",
+    "react-native-svg": "^9.3.6",
+    "react-native-thumbnail": "^1.1.3",
+    "react-native-typescript-transformer": "^1.2.10",
+    "react-native-view-overflow": "^0.0.4",
     "react-navigation": "^1.5.8",
-    "tslint-config-prettier": "^1.12.0"
+    "react-redux": "^5.1.1",
+    "redux": "^4.0.1",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0",
+    "tslint-config-prettier": "^1.12.0",
+    "tslint-react-recommended": "^1.0.15",
+    "yarn": "^1.12.3"
+  },
+  "rnpm": {
+    "assets": [
+      "./fonts/"
+    ]
   }
 }
