@@ -1,8 +1,10 @@
 import * as React from 'react'
-import { Image, View, Text, TouchableOpacity } from 'react-native'
-import { ImagePicker } from 'expo'
+import { View, TouchableOpacity } from 'react-native'
 import { StyleSheet } from 'react-native'
+
+import { RegularText, BoldText } from './TextAtom'
 import { color } from '../Style/Color'
+import CachedImageAtom from './CachedImageAtom'
 
 interface IProps {
   source: string
@@ -18,16 +20,7 @@ class ImageAtom extends React.Component<IProps, any> {
   }
 
   handleSelection = async () => {
-    if (this.props.getValue) {
-      let result = await ImagePicker.launchImageLibraryAsync({
-        allowsEditing: false
-      })
-
-      if (result && !result.cancelled) {
-        this.setState({ image: result })
-        this.props.getValue(this.state.image.uri)
-      }
-    }
+    // image picker functionality here
   }
 
   render() {
@@ -38,14 +31,12 @@ class ImageAtom extends React.Component<IProps, any> {
           style={styles.selfAlign}
         >
           <View style={styles.imgContainer}>
-            <Image
-              source={{ uri: this.state.image.uri || this.props.source }}
+            <CachedImageAtom
+              uri={this.state.image.uri || this.props.source}
               style={styles.image}
             />
           </View>
-          <Text style={[styles.imageText, { fontFamily: 'SourceSansPro' }]}>
-            Upload logo
-          </Text>
+          <RegularText style={[styles.imageText]}>Upload logo</RegularText>
         </TouchableOpacity>
       )
     } else {
@@ -55,12 +46,14 @@ class ImageAtom extends React.Component<IProps, any> {
           style={styles.selfAlign}
         >
           <View style={styles.imgContainer}>
-            <Text style={styles.imgPlaceholderText}>
+            <BoldText style={styles.imgPlaceholderText}>
               {this.props.placeholder &&
                 this.props.placeholder.substr(0, 1).toUpperCase()}
-            </Text>
+            </BoldText>
           </View>
-          <Text style={[styles.selfAlign, styles.menuColor]}>Upload logo</Text>
+          <RegularText style={[styles.selfAlign, styles.menuColor]}>
+            Upload logo
+          </RegularText>
         </TouchableOpacity>
       )
     }
@@ -92,7 +85,6 @@ const styles = StyleSheet.create({
     fontSize: 14
   },
   imgPlaceholderText: {
-    fontWeight: 'bold',
     fontSize: 20
   },
   menuColor: {
