@@ -3,15 +3,16 @@ import {
   View,
   ScrollView,
   StyleSheet,
-  Text,
   Dimensions,
   KeyboardAvoidingView
 } from 'react-native'
 import { Icon } from 'native-base'
-import InputAtom from '../Atom/InputAtom'
+
+import InputAtom from '../Atom/Form/InputAtom'
 import { color } from '../Style/Color'
 import SaveCancelButton from '../Container/SaveCancelButton'
 import FormInputViewAtom from '../Atom/FormInputViewAtom'
+import { RegularText } from '../Atom/TextAtom'
 
 interface IProps {
   navigation: any
@@ -63,7 +64,11 @@ class EmployeeForm extends Component<IProps, IState> {
               />
             </View>
             <View>
-              <Icon name="md-close" onPress={() => this.removeInput(key)} />
+              <Icon
+                name="md-close"
+                type="Ionicons"
+                onPress={() => this.removeInput(key)}
+              />
             </View>
           </View>
         </FormInputViewAtom>
@@ -102,14 +107,14 @@ class EmployeeForm extends Component<IProps, IState> {
                 />
               </View>
               <View>
-                <Icon name="md-close" />
+                <Icon name="md-close" type="Ionicons" />
               </View>
             </View>
           </FormInputViewAtom>
           {this.displayState()}
-          <Text style={styles.blueT} onPress={this.onPress}>
+          <RegularText style={styles.blueT} onPress={this.onPress}>
             +Add Email or Phone
-          </Text>
+          </RegularText>
         </ScrollView>
         <SaveCancelButton
           navigation={this.props.navigation}
@@ -151,6 +156,6 @@ const styles = StyleSheet.create({
     fontSize: 16,
     color: color.button,
     padding: 16,
-    fontFamily: 'SourceSansPro_Semibold'
+    fontFamily: 'SourceSansPro-Semibold'
   }
 })
