@@ -1,4 +1,4 @@
-export const numberWithCommas = (num: number) => {
+export const numberWithCommas = (num: number | string) => {
   let parts = num.toString().split('.')
   parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ',')
   return parts.join('.')
