@@ -2,31 +2,165 @@ import React, { PureComponent } from 'react'
 import {
   ScrollView,
   View,
-  Text,
+  Alert,
   TouchableOpacity,
   StyleSheet,
   Dimensions
 } from 'react-native'
 import { SafeAreaView } from 'react-navigation'
 import { Icon } from 'native-base'
+
 import { color } from '../Style/Color'
-import SideBarItemAtom from '../Atom/SideBarItemAtom'
 import Auth from '../services/auth'
+import { UserContext } from '../context/UserContext'
+import { MediumText } from '../Atom/TextAtom'
 
 interface IProps {
   navigation: any
+  onItemPress: any
+  items: any
+  gettingStartedProgress: any
 }
 
 interface IState {
   businessName: string
+  activeRoute: string
 }
 
-export default class SideBar extends PureComponent<IProps, IState> {
-  handleNavigation = (location: string, data: any = undefined) => {
-    this.props.navigation.navigate(location, { data })
+interface Category {
+  title: string
+  iconName?: string
+  iconType?: any
+  routeName: string
+}
+
+const sideBarItemStyles = StyleSheet.create({
+  categoryContainer: {
+    borderBottomWidth: StyleSheet.hairlineWidth,
+    borderBottomColor: '#d8d8d8',
+    paddingVertical: 8
+  },
+  hideBorder: {
+    borderBottomWidth: 0,
+    paddingVertical: 6
+  },
+  listHeader: {
+    backgroundColor: 'transparent',
+    marginTop: 16,
+    marginBottom: 5
+  },
+  title: {
+    marginLeft: 25,
+    color: '#999',
+    marginVertical: 2,
+    fontSize: 11
+  },
+  categoryWrapper: {
+    borderRadius: 5,
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginHorizontal: 5,
+    paddingLeft: 22,
+    height: 48
+  },
+  category: {
+    marginLeft: 20,
+    backgroundColor: 'transparent',
+    color: color.textColor,
+    fontSize: 14
+  },
+  activeCategoryWrapper: {
+    backgroundColor: '#e1f5fe'
+  },
+  activeCategory: {
+    color: color.button
   }
+})
+
+const SidebarItem = (prop: {
+  title: string
+  categories: Category[]
+  navigate: any
+  hideBottomBorder?: boolean
+  activeRoute?: string | null
+}) => {
+  return (
+    <View
+      style={
+        prop.hideBottomBorder
+          ? sideBarItemStyles.hideBorder
+          : sideBarItemStyles.categoryContainer
+      }
+    >
+      {prop.categories.map((category: Category, key: number) => {
+        return (
+          <TouchableOpacity
+            style={[
+              sideBarItemStyles.categoryWrapper,
+              prop.activeRoute == category.routeName &&
+                sideBarItemStyles.activeCategoryWrapper
+            ]}
+            key={key}
+            onPress={() => prop.navigate(category.routeName)}
+          >
+            <View style={{ flexDirection: 'row', alignItems: 'center' }}>
+              <View style={{ width: 20 }}>
+                <Icon
+                  name={category.iconName}
+                  type={category.iconType}
+                  style={[
+                    {
+                      fontSize: category.iconType == 'EvilIcons' ? 27 : 20,
+                      marginLeft: category.iconType == 'EvilIcons' ? -5 : 0,
+                      color: color.textColor
+                    },
+                    prop.activeRoute == category.routeName &&
+                      sideBarItemStyles.activeCategory
+                  ]}
+                />
+              </View>
+              <MediumText
+                style={[
+                  sideBarItemStyles.category,
+                  prop.activeRoute == category.routeName &&
+                    sideBarItemStyles.activeCategory
+                ]}
+              >
+                {category.title}
+              </MediumText>
+            </View>
+          </TouchableOpacity>
+        )
+      })}
+    </View>
+  )
+}
+
+class SideBar extends PureComponent<IProps, IState> {
+  handleNavigation = async (location: string) => {
+    this.setState({ activeRoute: location })
+    this.props.gettingStartedProgress == 'done'
+      ? this.props.navigation.navigate(location)
+      : this.showAlert()
+  }
+
   state = {
-    businessName: ''
+    businessName: '',
+    activeRoute: ''
+  }
+
+  showAlert = () => {
+    Alert.alert(
+      'Sorry!!!',
+      `We know you can't wait to begin, but you have to complete the get started section to do so.`,
+      [
+        {
+          text: 'Ok',
+          onPress: () => null
+        }
+      ],
+      { cancelable: false }
+    )
   }
 
   componentWillMount() {
@@ -54,96 +188,132 @@ export default class SideBar extends PureComponent<IProps, IState> {
               style={styles.header}
               onPress={() => navigate('DrawerToggle')}
             >
-              <Icon name="cross" type="Entypo" style={styles.cross} />
-              <Text style={[styles.texts, { fontFamily: 'SourceSansPro' }]}>
+              <Icon name="md-close" type="Ionicons" style={styles.cross} />
+              {/* <Text style={[styles.texts, { fontFamily: 'Source Sans Pro' }]}>
                 {this.state.businessName}
-              </Text>
+              </Text> */}
             </TouchableOpacity>
-            <SideBarItemAtom
+            <SidebarItem
               title="COMPANY"
-              navigate={navigate}
+              navigate={this.handleNavigation}
+              activeRoute={this.state.activeRoute}
               categories={[
                 {
                   title: 'Home',
+                  iconName: 'home',
+                  iconType: 'SimpleLineIcons',
                   routeName: 'Home'
-                },
-                {
-                  title: 'Products & Services',
-                  routeName: 'ViewBusiness'
-                },
-                {
-                  title: 'Employees',
-                  routeName: 'Employees'
                 }
               ]}
             />
 
-            <SideBarItemAtom
+            <SidebarItem
+              title="STORE"
+              navigate={this.handleNavigation}
+              activeRoute={this.state.activeRoute}
+              categories={[
+                {
+                  title: 'Products',
+                  routeName: 'Products',
+                  iconName: 'archive',
+                  iconType: 'EvilIcons'
+                },
+                {
+                  title: 'Categories',
+                  routeName: 'Categories',
+                  iconName: 'folder-alt',
+                  iconType: 'SimpleLineIcons'
+                },
+                {
+                  title: 'Price Calculator',
+                  routeName: 'ChargeCalculator',
+                  iconName: 'calculator',
+                  iconType: 'SimpleLineIcons'
+                }
+              ]}
+            />
+
+            <SidebarItem
               title="CONTACT"
-              navigate={navigate}
+              navigate={this.handleNavigation}
+              activeRoute={this.state.activeRoute}
               categories={[
                 {
-                  title: 'Customers',
-                  routeName: 'Customers'
-                },
-                {
-                  title: 'Vendors',
-                  routeName: 'Vendors'
-                }
-              ]}
-            />
-
-            <SideBarItemAtom
-              title="ORDER"
-              navigate={navigate}
-              categories={[
-                {
-                  title: 'Purchase',
-                  routeName: 'Purchase'
+                  title: 'Contacts',
+                  routeName: 'Contacts',
+                  iconName: 'user-o',
+                  iconType: 'FontAwesome'
                 },
                 {
                   title: 'Sales',
-                  routeName: 'salesOrderTab'
+                  routeName: 'Sales',
+                  iconName: 'cart-outline',
+                  iconType: 'MaterialCommunityIcons'
+                },
+                {
+                  title: 'Shipping Zones',
+                  routeName: 'DeliveryFees',
+                  iconName: 'location-pin',
+                  iconType: 'SimpleLineIcons'
                 }
               ]}
             />
-            <SideBarItemAtom
-              title="TRANSACTIONS"
-              navigate={navigate}
+
+            <SidebarItem
+              title="ORDER"
+              navigate={this.handleNavigation}
+              activeRoute={this.state.activeRoute}
               categories={[
                 {
-                  title: 'Bank',
-                  routeName: 'Bank'
-                },
-                {
-                  title: 'Income',
-                  routeName: 'Income'
+                  title: 'Invoices',
+                  routeName: 'Invoices',
+                  iconName: 'receipt',
+                  iconType: 'MaterialCommunityIcons'
                 },
                 {
                   title: 'Expenses',
-                  routeName: 'Expenses'
+                  routeName: 'Expenses',
+                  iconName: 'credit-card',
+                  iconType: 'SimpleLineIcons'
                 },
                 {
-                  title: 'Invoices',
-                  routeName: 'Invoice'
-                },
-                {
-                  title: 'Receipts',
-                  routeName: 'Receipts'
+                  title: 'Banks',
+                  routeName: 'Banks',
+                  iconName: 'bank',
+                  iconType: 'MaterialCommunityIcons'
                 }
               ]}
             />
-            <SideBarItemAtom
+            {/* <SidebarItem
+              title="OFFERS"
+              navigate={this.handleNavigation}
+              activeRoute={this.state.activeRoute}
+              categories={[
+                {
+                  title: 'Special Offers',
+                  routeName: 'SpecialOffer',
+                  iconName: 'local-offer',
+                  iconType: 'MaterialIcons'
+                }
+              ]}
+            /> */}
+            <SidebarItem
               title="HELP & SETTINGS"
-              navigate={navigate}
+              hideBottomBorder
+              navigate={this.handleNavigation}
+              activeRoute={this.state.activeRoute}
               categories={[
                 {
                   title: 'Settings',
-                  routeName: 'ProfileSettings'
+                  routeName: 'ProfileSettings',
+                  iconType: 'SimpleLineIcons',
+                  iconName: 'settings'
                 },
                 {
                   title: 'Help & Feedback',
-                  routeName: 'Help'
+                  routeName: 'Help',
+                  iconType: 'Ionicons',
+                  iconName: 'ios-help-circle-outline'
                 }
               ]}
             />
@@ -154,12 +324,22 @@ export default class SideBar extends PureComponent<IProps, IState> {
   }
 }
 
+const _SideBar = props => (
+  <UserContext.Consumer>
+    {({ gettingStartedProgress }) => (
+      <SideBar {...props} gettingStartedProgress={gettingStartedProgress} />
+    )}
+  </UserContext.Consumer>
+)
+
+export default _SideBar
+
 const styles = StyleSheet.create({
   header: {
     flexDirection: 'row',
-    marginLeft: '3%',
-    alignItems: 'center',
-    paddingVertical: 16
+    marginLeft: 27,
+    alignItems: 'center'
+    // paddingVertical: 16
   },
   texts: {
     color: color.modal,
@@ -176,9 +356,8 @@ const styles = StyleSheet.create({
     color: color.menu
   },
   cross: {
-    fontSize: 35,
-    // backgroundColor: 'transparent',
-    color: color.modal
+    fontSize: 26,
+    color: '#000'
   },
   logoutItem: {
     borderTopWidth: 1,
@@ -192,7 +371,9 @@ const styles = StyleSheet.create({
   },
   sidebarContainer: {
     height: Dimensions.get('window').height - 16,
-    backgroundColor: color.primary
+    backgroundColor: '#fff',
+    paddingTop: 20,
+    paddingBottom: 30
   },
   itemsContainer: {
     flex: 4
