@@ -1,7 +1,7 @@
-import gql from "graphql-tag";
+import gql from 'graphql-tag'
 
 export const AuthenticateClientGQL = gql`
-  mutation authenticate {
-    authenticate @client
+  mutation authenticate($user: User!) {
+    authenticate(user: $user) @client
   }
-`;
+`
