@@ -1,4 +1,4 @@
-import gql from 'graphql-tag';
+import gql from 'graphql-tag'
 
 export const LoginUserMutationGQL = gql`
   mutation loginUser($email: String!, $password: String!) {
@@ -12,6 +12,10 @@ export const LoginUserMutationGQL = gql`
         ... on Authorization {
           message
           accessToken
+          s3Bucket
+          s3Region
+          s3AccessKey
+          s3SecretKey
           refreshToken
           user {
             id
@@ -21,25 +25,42 @@ export const LoginUserMutationGQL = gql`
             dateOfBirth
             gender
             profilePicture
-            phone {
-              type
-              number
-            }
-            location {
-              id
-              city
-              country
-              state
-              street1
-              type
-            }
             company {
               id
               title
+              saleCharge
+              shareLink
               contactEmail
               about
-              category
               currency
+              deliveryFees {
+                id
+                state
+                region
+                fee
+              }
+              bank {
+                accountNumber
+                bankCode
+                subaccountId
+                subaccountTransacId
+              }
+              slug
+              logo
+              facebook
+              twitter
+              legalDocuments {
+                pdfUrl
+                name
+                type
+                id
+              }
+              instagram
+              linkedin
+              coverPhoto
+              phone {
+                number
+              }
               branches {
                 id
                 type
@@ -58,46 +79,90 @@ export const LoginUserMutationGQL = gql`
       }
     }
   }
-`;
-
-export const RegisterCompanyMutationGQL = gql`
-  mutation registerCompany(
-    $firstName: String!
-    $lastName: String!
-    $password: String!
-    $passwordConfirmation: String!
-    $contactEmail: String!
-    $currency: String
-    $street1: String!
-    $city: String!
-    $state: String!
-    $country: String!
-    $category: Category!
-    $title: String!
-    $email: String!
-  ) {
-    registerCompany(
-      company: {
-        category: $category
-        contactEmail: $contactEmail
-        currency: $currency
-        headOffice: {
-          street1: $street1
-          city: $city
-          state: $state
-          country: $country
+`
+export const RegisterUserMutationGQL = gql`
+  mutation registerUser($user: UserInput!) {
+    registerUser(user: $user) {
+      success
+      fieldErrors {
+        key
+        message
+      }
+      data {
+        ... on Authorization {
+          message
+          accessToken
+          s3Bucket
+          s3Region
+          s3AccessKey
+          s3SecretKey
+          refreshToken
+          user {
+            id
+            email
+            firstName
+            lastName
+            dateOfBirth
+            gender
+            profilePicture
+            company {
+              id
+              saleCharge
+              title
+              slug
+              shareLink
+              contactEmail
+              about
+              deliveryFees {
+                id
+                state
+                region
+                fee
+              }
+              legalDocuments {
+                pdfUrl
+                name
+                type
+                id
+              }
+              bank {
+                accountNumber
+                bankCode
+                subaccountId
+                subaccountTransacId
+              }
+              currency
+              coverPhoto
+              facebook
+              twitter
+              instagram
+              linkedin
+              logo
+              phone {
+                number
+              }
+              branches {
+                id
+                type
+                location {
+                  id
+                  city
+                  country
+                  state
+                  street1
+                  type
+                }
+              }
+            }
+          }
         }
-        title: $title
       }
-      user: {
-        email: $email
-        firstName: $firstName
-        gender: MALE
-        lastName: $lastName
-        password: $password
-        passwordConfirmation: $passwordConfirmation
-      }
-    ) {
+    }
+  }
+`
+export const AddUserCompanyMutationGQL = gql`
+  mutation addUserCompany($company: CompanyInput!, $userId: Uuid!) {
+    addUserCompany(company: $company, user: $userId) {
       fieldErrors {
         key
         message
@@ -105,10 +170,49 @@ export const RegisterCompanyMutationGQL = gql`
       success
       data {
         ... on Company {
+          saleCharge
           id
           title
+          slug
+          contactEmail
+          shareLink
+          about
+          currency
+          deliveryFees {
+            id
+            state
+            region
+            fee
+          }
+          facebook
+          legalDocuments {
+            pdfUrl
+            name
+            type
+            id
+          }
+          twitter
+          instagram
+          linkedin
+          coverPhoto
+          logo
+          phone {
+            number
+          }
+          branches {
+            id
+            type
+            location {
+              id
+              city
+              country
+              state
+              street1
+              type
+            }
+          }
         }
       }
     }
   }
-`;
+`
