@@ -1,50 +1,8 @@
-import gql from 'graphql-tag';
+import gql from 'graphql-tag'
 
 export const UpsertContactGQL = gql`
-  mutation upsertContact(
-    $contactId: Uuid
-    $street1: String!
-    $city: String!
-    $state: String!
-    $country: String!
-    $dislikes: [String]
-    $likes: [String]
-    $contactName: String!
-    $currency: String
-    $birthday: String
-    $maritalStatus: String
-    $number: String!
-    $userId: Uuid!
-    $image: String
-    $email: String!
-    $bank: BankInput
-    $companyId: Uuid!
-    $type: String!
-  ) {
-    upsertContact(
-      contact: {
-        userId: $userId
-        email: $email
-        likes: $likes
-        dislikes: $dislikes
-        birthday: $birthday
-        currency: $currency
-        contactName: $contactName
-        bank: $bank
-        maritalStatus: $maritalStatus
-        image: $image
-        phone: { number: $number }
-        type: $type
-        address: {
-          state: $state
-          street1: $street1
-          city: $city
-          country: $country
-        }
-        companyId: $companyId
-      }
-      contactId: $contactId
-    ) {
+  mutation upsertContact($contactId: Uuid, $contact: ContactInput!) {
+    upsertContact(contact: $contact, contactId: $contactId) {
       success
       fieldErrors {
         key
@@ -54,8 +12,11 @@ export const UpsertContactGQL = gql`
         ... on Contact {
           id
           contactName
+          allowsMarketing
           email
           image
+          type
+          gender
           address {
             state
             street1
@@ -73,14 +34,28 @@ export const UpsertContactGQL = gql`
           maritalStatus
           currency
           birthday
-          bank {
-            bankName
-            accountName
-            accountNumber
-            id
-          }
+
+          instagram
+          facebook
+          twitter
+          snapchat
+          totalDebt
+          totalAmountPaid
+
+          data: updatedAt
         }
       }
     }
   }
-`;
+`
+export const DeleteContact = gql`
+  mutation deleteContact($contactId: Uuid!) {
+    deleteContact(contactId: $contactId) {
+      success
+      fieldErrors {
+        key
+        message
+      }
+    }
+  }
+`
